requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
